BTW CE 1.5.2 port

+++ src/main/java/makamys/neodymium/util/GuiHelper.java
@@ -1,9 +1,9 @@
 package makamys.neodymium.util;
 
+import net.minecraft.src.Tessellator;
 import org.lwjgl.opengl.GL11;
 
 import net.minecraft.client.Minecraft;
-import net.minecraft.client.renderer.Tessellator;
 
 public class GuiHelper {
     
@@ -16,7 +16,7 @@ public class GuiHelper {
         GL11.glMatrixMode(GL11.GL_PROJECTION);
         //GL11.glEnable(GL11.GL_COLOR_MATERIAL);
         GL11.glLoadIdentity();
-        GL11.glOrtho(0.0D, (double)mc.displayWidth, (double)mc.displayHeight, 0.0D, 1000.0D, 3000.0D);
+        GL11.glOrtho(0.0D, mc.displayWidth, mc.displayHeight, 0.0D, 1000.0D, 3000.0D);
         GL11.glMatrixMode(GL11.GL_MODELVIEW);
         GL11.glLoadIdentity();
         GL11.glTranslatef(0.0F, 0.0F, -2000.0F);
